refactor(movements): Move movements to Movements namespace
